Remove mocked PSR7 Response objects
Fixes an issue where the mock was returning the wrong type: Psr7\Response::getBody returns an instance of StreamInterface, not a string, and with the move to Guzzle 7 the type difference is causing problems.

The solution was to stop mocking the response altogether; it doesn't really test anything, and using a real Response object with expected status and body means that subsequent calls to getBody() correctly return instances of StreamInterface.

Also - this was the only place we were mocking this way, across various repos, so it's now in line