DDLS-688 Upgrade to PHP 8.4 / remove Encore (#737)
* Upgrade Dockerfiles and composer config to PHP 8.4.8

* Application runs, most functionality works
* Unit tests pass, but with lots of deprecations
* User add doesn't work, but I imagine that doesn't locally anyway
(notify critical error, and I can't see where the notify mock which
is in this repo runs)

* escape must be explicitly passed to fputcsv for behaviour to be retained

Not passing an escape parameter causes numerous deprecation warnings.

* Fix deprecations in OrderController test

* Fix various deprecations in tests

* Streamline JS into a single file

* Fix issues with behat tests

* Directory not available caused error message
* Deprecations make behat tests fail

* Use PHP 8.4.8 to run behat tests

* Use esbuild for JS and sass for CSS

* Add static assets (not complete)

This copies govuk images and fonts to the distribution.

However, the icon paths are incorrect, and icons won't load
via references in the CSS. This seems to be mostly an
issue with how the dropzone CSS refers to icons and is
probably fixable.

* Repair Dropzone

* Fix icon paths

* Remove leftovers from webpack and encore

* Add some documentation to the build script

* Clean up ignore file

* Use standard ordering for imports and constants